Adds CPU configs, GIN configs and better modularization, update to dependencies and README
