+++ env.yml
@@ -71,3 +71,4 @@ dependencies:
 
   - pip:
       - lightning-graphcore # optional, for using IPUs only
+      - hydra-core>=1.3.2
