+++ expts/hydra-configs/accelerator/cpu.yaml
@@ -1,0 +1,1 @@
+type: cpu
