+++ expts/hydra-configs/dataset/accelerator/toymix_gpu.yaml
