+++ expts/hydra-configs/experiment/toymix_gin.yaml
@@ -1,13 +1,12 @@
 # @package _global_
 
 constants:
-  name: neurips2023_small_data_gcn
+  name: neurips2023_small_data_gin
   entity: "multitask-gnn"
   seed: 42
-  max_epochs: 100
   data_dir: expts/data/neurips2023/small-dataset
   raise_train_error: true
 
 trainer:
   model_checkpoint:
-    dirpath: models_checkpoints/neurips2023-small-gcn/
+    dirpath: models_checkpoints/neurips2023-small-gin/
