+++ expts/hydra-configs/model/gcn.yaml
@@ -1,0 +1,5 @@
+# @package _global_
+
+architecture:
+  gnn:
+    layer_type: 'pyg:gcn'
