+++ expts/main_run_multitask.py
@@ -1,10 +1,7 @@
 # General imports
-import argparse
 import os
 from os.path import dirname, abspath
-import yaml
-from copy import deepcopy
-from omegaconf import DictConfig
+from omegaconf import DictConfig, OmegaConf
 import timeit
 from loguru import logger
 from datetime import datetime
@@ -20,43 +17,33 @@ from graphium.config._loader import (
     load_trainer,
     save_params_to_wandb,
     load_accelerator,
-    load_yaml_config,
 )
 from graphium.utils.safe_run import SafeRun
-from graphium.utils.command_line_utils import update_config, get_anchors_and_aliases
 
+import hydra
 
 # WandB
 import wandb
 
 # Set up the working directory
 MAIN_DIR = dirname(dirname(abspath(graphium.__file__)))
-
-# CONFIG_FILE = "expts/configs/config_mpnn_10M_b3lyp.yaml"
-# CONFIG_FILE = "expts/configs/config_mpnn_10M_pcqm4m.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_debug.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_large_mpnn.yaml"
-CONFIG_FILE = "expts/neurips2023_configs/debug/config_small_gcn_debug.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_large_gin.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_large_gcn.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_large_gine.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_small_gcn.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_small_gin.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_small_gine.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_small_gcn_gpu.yaml"
-# CONFIG_FILE = "expts/neurips2023_configs/config_large_gcn_gpu.yaml"
-
 os.chdir(MAIN_DIR)
 
 
-def main(cfg: dict, run_name: str = "main", add_date_time: bool = True) -> None:
+@hydra.main(version_base=None, config_path="hydra-configs", config_name="main")
+def main(cfg: DictConfig) -> None:
+    cfg = OmegaConf.to_container(cfg, resolve=True)
+
+    run_name: str = "main"
+    add_date_time: bool = True
+
     st = timeit.default_timer()
 
     date_time_suffix = ""
     if add_date_time:
         date_time_suffix = datetime.now().strftime("%d.%m.%Y_%H.%M.%S")
 
-    wandb.init(entity="multitask-gnn", project=cfg["constants"]["name"], config=cfg)
+    wandb.init(entity=cfg["constants"]["entity"], project=cfg["constants"]["name"], config=cfg)
 
     # Initialize the accelerator
     cfg, accelerator_type = load_accelerator(cfg)
@@ -109,12 +96,4 @@ def main(cfg: dict, run_name: str = "main", add_date_time: bool = True) -> None:
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--config", help="Path to the config file", default=None)
-
-    args, unknown_args = parser.parse_known_args()
-    if args.config is not None:
-        CONFIG_FILE = args.config
-    cfg = load_yaml_config(CONFIG_FILE, MAIN_DIR, unknown_args)
-
-    main(cfg)
+    main()
